requirements.txt
langchain>=0.3.7
langchain-community>=0.3.7
langchain-core>=0.3.15
langchain-google-genai>=2.0.5
langchain-postgres>=0.0.12
langchain-text-splitters>=0.3.2
pypdf>=5.1.0
python-dotenv>=1.0.1
psycopg[binary]>=3.2.3
